Adding JS test setup
