+++ gulpfile.js
@@ -17,6 +17,10 @@ var path = require('path');
 var livereload = require('gulp-livereload');
 var modernizr = require('gulp-modernizr');
 var fileExists = require('file-exists');
+var mocha = require('gulp-mocha');
+var istanbul = require('gulp-istanbul');
+var isparta = require('isparta');
+var coverageEnforcer = require('gulp-istanbul-enforcer');
 
 var spawn = require('child_process').spawn;
 var argv = require('yargs')
@@ -180,6 +184,50 @@ gulp.task('default', function (cb) {
     }
   });
 
+});
+
+gulp.task('test', function () {
+  require('babel-core/register');
+  return gulp
+    .src('./traffic_stops/static/js/app/**/*.js')
+    .pipe(istanbul({
+      instrumenter: isparta.Instrumenter
+      , includeUntested: true
+    }))
+    .pipe(istanbul.hookRequire())
+    .on('finish', function () {
+      gulp
+        .src('./traffic_stops/static/js/test/**/test_*.js', {read: false})
+        .pipe(mocha({
+          require: [
+            'jsdom-global/register'
+          ]
+        }))
+        .pipe(istanbul.writeReports({
+          dir: './coverage/'
+          , reportOpts: {
+            dir: './coverage/'
+          }
+          , reporters: [
+            'text'
+            , 'text-summary'
+            , 'json'
+            , 'html'
+          ]
+        }))
+        .pipe(coverageEnforcer({
+          thresholds: {
+            statements: 80
+            , branches: 50
+            , lines: 90
+            , functions: 50
+          }
+          , coverageDirectory: './coverage/'
+          , rootDirectory: ''
+        }))
+      ;
+    })
+  ;
 });
 
 gulp.task('deploy', function() {
