+++ package.json
@@ -27,6 +27,7 @@
   "dependencies": {},
   "devDependencies": {
     "@caktus/flexless": "^0.1.1",
+    "babel-core": "^6.10.4",
     "babel-preset-es2015": "^6.1.18",
     "babelify": "^7.2.0",
     "backbone": "1.1.2",
@@ -40,17 +41,24 @@
     "gulp-connect": "^2.2.0",
     "gulp-cssmin": "^0.1.6",
     "gulp-if": "^1.2.4",
+    "gulp-istanbul": "^1.0.0",
+    "gulp-istanbul-enforcer": "^1.0.3",
     "gulp-jasmine2-phantomjs": "^0.1.1",
     "gulp-less": "^3.0.5",
     "gulp-livereload": "^2.1.1",
+    "gulp-mocha": "^2.2.0",
     "gulp-modernizr": "0.0.0",
     "gulp-notify": "^1.4.2",
     "gulp-rename": "^1.2.2",
     "gulp-streamify": "0.0.5",
     "gulp-uglify": "^0.3.1",
     "gulp-util": "^3.0.0",
+    "isparta": "^4.0.0",
+    "istanbul": "^0.4.4",
     "jquery": "^2.1",
     "less": "^2.5.3",
+    "mocha": "^2.5.3",
+    "mocha-jsdom": "^1.1.0",
     "nvd3": "1.1.15",
     "tabslet": "^1.7.2",
     "underscore": "1.7.0",
