+++ traffic_stops/static/js/app/states/md/index.js
@@ -1,0 +1,1 @@
+import './Stops.js';
