--- traffic_stops/static/js/states/md/index.js
@@ -1,1 +1,0 @@
-import './Stops.js';
