--- traffic_stops/static/js/states/nc/Stops.js
@@ -1,248 +1,0 @@
-import DataHandlerBase from '../../base/DataHandlerBase.js';
-import VisualBase from '../../base/VisualBase.js';
-import TableBase from '../../base/TableBase.js';
-import Stops from './defaults.js';
-
-import Backbone from 'backbone';
-import _ from 'underscore';
-import d3 from 'd3';
-import $ from 'jquery';
-Backbone.$ = $;
-
-export var StopsHandler = DataHandlerBase.extend({
-  clean_data: function(){
-
-    var data = this.get("raw_data"),
-        total = {};
-
-    // build a "Totals" year which sums by race/ethnicity for all years
-    if (data.length>0){
-      // create new totals object, and reset values
-      total = _.clone(data[0]);
-      _.keys(total).forEach(function(key){
-        total[key] = 0;
-      });
-
-      // sum data from all years
-      data.forEach(function(year){
-        _.keys(year).forEach(function(key){
-          total[key] += year[key];
-        });
-      });
-      total["year"] = "Total";
-    }
-
-    // build-data for pie-chart
-    var pie = d3.map();
-    data.forEach(function(v){
-      if (v.year>=Stops.start_year) pie.set(v.year, d3.map(v));
-    });
-    pie.set("Total", d3.map(total));
-
-    // build data for line-chart
-
-    var line = d3.map(),
-        get_total_by_race = function(dataType, yr){
-          var total = 0;
-          dataType.forEach(function(race){
-            total += yr[race];
-          });
-          return total;
-        };
-
-    [Stops.races, Stops.ethnicities].forEach(function(dataType){
-      dataType.forEach(function(v){
-        line.set(v, []);
-      });
-      data.forEach(function(yr){
-        if (yr.year>=Stops.start_year){
-          var total = get_total_by_race(dataType, yr);
-          dataType.forEach(function(race){
-            line.get(race).push({x: yr.year, y:yr[race]/total});
-          });
-        }
-      });
-    });
-
-    // set object data
-    this.set("data", {
-      type: "stop",
-      raw: this.get("raw_data"),
-      pie: pie,
-      line: line
-    });
-  }
-});
-
-export var StopRatioDonut = VisualBase.extend({
-  defaults: {
-    showEthnicity: false,
-    width: 300,
-    height: 300
-  },
-  setDefaultChart: function(){
-    this.chart = nv.models.pieChart()
-      .x(function(d){ return d.key; })
-      .y(function(d){ return d.value; })
-      .color(function(d){ return d.data.color; })
-      .width(this.get("width"))
-      .height(this.get("height"))
-      .showLabels(true)
-      .labelType("percent")
-      .donutRatio(0.35)
-      .labelThreshold(0.05)
-      .donut(true);
-  },
-  drawStartup: function(){
-
-    // get year options for pulldown menu
-    var selector = $('<select>'),
-        year_options = this.data.pie.keys(),
-        opts = year_options.map((v) => `<option value="${v}">${v}</option>`),
-        getData = () => {
-          var value = selector.val();
-          this.dataset =  this.data.pie.get(value);
-          this.drawChart();
-        };
-
-    selector
-      .append(opts)
-      .val("Total")
-      .on('change', getData);
-
-    $('<div>')
-      .html(selector)
-      .appendTo(this.div);
-
-    getData();
-  },
-  drawChart: function(){
-    var data = this._formatData();
-
-    nv.addGraph(() => {
-      d3.select(this.svg[0])
-          .datum(data)
-        .transition().duration(1200)
-          .attr('width', "100%")
-          .attr('height', "100%")
-          .attr("preserveAspectRatio", "xMinYMin")
-          .attr('viewBox', `0 0 ${this.get('width')} ${this.get('height')}`)
-          .call(this.chart);
-    });
-  },
-  _formatData: function(){
-    var data = [],
-        selected = this.dataset,
-        items = (this.get('showEthnicity')) ? Stops.ethnicities : Stops.races;
-
-    // build data specifically for this pie chart
-    items.forEach(function(d, i){
-      data.push({
-        "key": Stops.pprint.get(d),
-        "value": selected.get(d),
-        "color": Stops.colors[i]
-      });
-    });
-
-    return data;
-  },
-  triggerRaceToggle: function(e, v){
-    this.set('showEthnicity', v);
-    this.drawChart();
-  }
-});
-
-export var StopRatioTimeSeries = VisualBase.extend({
-  defaults: {
-    showEthnicity: false,
-    width: 750,
-    height: 375
-  },
-  setDefaultChart: function(){
-    this.chart = nv.models.lineChart()
-                  .useInteractiveGuideline (true)
-                  .transitionDuration(350)
-                  .showLegend(true)
-                  .showYAxis(true)
-                  .showXAxis(true)
-                  .forceY([0, 1])
-                  .width(this.get("width"))
-                  .height(this.get("height"));
-
-    this.chart.xAxis
-        .axisLabel('Year');
-
-    this.chart.yAxis
-        .axisLabel('Percentage of stops by race')
-        .tickFormat(d3.format('%'));
-  },
-  drawStartup: function(){},
-  drawChart: function(){
-    var data = this._formatData();
-
-    nv.addGraph(() => {
-        d3.select(this.svg[0])
-          .datum(data)
-          .attr('width', "100%")
-          .attr('height', "100%")
-          .attr('preserveAspectRatio', "xMinYMin")
-          .attr('viewBox', `0 0 ${this.get('width')} ${this.get('height')}`)
-          .call(this.chart);
-      });
-  },
-  _formatData: function(){
-    var data = [],
-        items = (this.get('showEthnicity')) ? Stops.ethnicities : Stops.races,
-        subset = [],
-        i = 0,
-        disabled;
-
-    this.data.line.forEach(function(key, vals){
-      if (items.indexOf(key) < 0) return;
-      // disable by default if maximum value < 5%
-      disabled = d3.max(vals, function(d){return d.y;})<0.05;
-      data.push({
-        key: Stops.pprint.get(key),
-        values: vals,
-        color: Stops.colors[i],
-        disabled: disabled
-      });
-      i += 1;
-    });
-    return data;
-  },
-  triggerRaceToggle: function(e, v){
-    this.set('showEthnicity', v);
-    this.drawChart();
-  }
-});
-
-export var StopsTable = TableBase.extend({
-  get_tabular_data: function(){
-    var header, row, rows = [];
-
-    // create header
-    header = ["Year"];
-    header.push.apply(header, Stops.pprint.values());
-    rows.push(header);
-
-    // create data rows
-    this.data.pie.forEach(function(k, v){
-      row = [k];
-      Stops.races.forEach(function(r){ row.push((v.get(r)||0).toLocaleString()); });
-      Stops.ethnicities.forEach(function(e){ row.push((v.get(e)||0).toLocaleString()); });
-      rows.push(row);
-    });
-
-    return rows;
-  }
-});
-
-if (typeof window.NC === 'undefined') window.NC = {};
-
-Object.assign(window.NC, {
-  StopsHandler,
-  StopRatioDonut,
-  StopRatioTimeSeries,
-  StopsTable
-});
